LayoutTests/svg/animations/getCurrentTime-pause-unpause-expected.txt

PASS svg.getCurrentTime() is 0
PASS svg.getCurrentTime() is 0.05
PASS svg.getCurrentTime() is 0.05
PASS successfullyParsed is true

TEST COMPLETE
